feat(dedicated): lazy load otrs section
