--- packages/manager/apps/dedicated/client/app/account/otrs/otrs-controllers.module.js
@@ -1,1 +1,0 @@
-angular.module('Module.otrs.controllers', []);
