--- packages/manager/apps/dedicated/client/app/account/otrs/otrs-filters.module.js
@@ -1,1 +1,0 @@
-angular.module('Module.otrs.filters', []);
