+++ packages/manager/apps/dedicated/client/app/app.js
@@ -1,5 +1,4 @@
 import { Environment } from '@ovh-ux/manager-config';
-import { buildURL } from '@ovh-ux/ufrontend/url-builder';
 import get from 'lodash/get';
 import has from 'lodash/has';
 import set from 'lodash/set';
@@ -12,7 +11,6 @@ import ngOvhBrowserAlert from '@ovh-ux/ng-ovh-browser-alert';
 import ngOvhExportCsv from '@ovh-ux/ng-ovh-export-csv';
 import ngOvhFeatureFlipping from '@ovh-ux/ng-ovh-feature-flipping';
 import ngOvhHttp from '@ovh-ux/ng-ovh-http';
-import ngOvhOtrs from '@ovh-ux/ng-ovh-otrs';
 import ngOvhProxyRequest from '@ovh-ux/ng-ovh-proxy-request';
 import ngOvhSsoAuth from '@ovh-ux/ng-ovh-sso-auth';
 import ngOvhSsoAuthModalPlugin from '@ovh-ux/ng-ovh-sso-auth-modal-plugin';
@@ -66,6 +64,7 @@ import expiredPage from './expired';
 
 import dedicatedServer from './dedicated/server';
 import userContracts from './user-contracts';
+import otrs from './otrs';
 
 import { TRACKING } from './at-internet.constants';
 
@@ -98,7 +97,7 @@ angular
       managedBaremetal,
       'Module.ip',
       'Module.license',
-      'Module.otrs',
+      otrs,
       ovhManagerMfaEnrollment,
       'ng.ckeditor',
       'ngMessages',
@@ -108,7 +107,6 @@ angular
       ngOvhBrowserAlert,
       ngOvhFeatureFlipping,
       ngOvhHttp,
-      ngOvhOtrs,
       ngOvhProxyRequest,
       ngOvhSsoAuth,
       ngOvhSsoAuthModalPlugin,
@@ -326,9 +324,6 @@ angular
   })
   .config(($qProvider) => {
     $qProvider.errorOnUnhandledRejections(false);
-  })
-  .config((OtrsPopupProvider) => {
-    OtrsPopupProvider.setBaseUrlTickets(buildURL('dedicated', '#/ticket'));
   })
   .run(
     /* @ngInject */ ($translate) => {
