+++ packages/manager/apps/dedicated/client/app/app.module.js
@@ -62,10 +62,6 @@ import moduleName from './app';
 import './app.routes';
 
 import './account/contacts/service/edit/edit.module';
-import './account/otrs/otrs-controllers.module';
-import './account/otrs/otrs-directives.module';
-import './account/otrs/otrs-filters.module';
-import './account/otrs/otrs-services.module';
 import './account/user/advanced/advanced.module';
 import './account/user/dashboard/user-dahboard.module';
 import './account/user/newAccountForm/new-account-form.module';
