+++ packages/manager/apps/dedicated/client/app/index.html
@@ -131,7 +131,7 @@
             data-support-level="{{$ctrl.user.supportLevel.level}}"
         ></ovh-chatbot>
 
-        <div data-otrs-popup></div>
+        <div id="otrs-popup" data-otrs-popup></div>
         <script>
             var CKEDITOR_BASEPATH =
                 (window.location.pathname !== '/'
