+++ packages/manager/apps/dedicated/client/app/otrs/detail/otrs-detail.html
